refactor: Improve proof of work implementation in Block
This commit refactors the `Block` struct to improve the implementation of the proof of work algorithm. The `run_proof_of_work` method has been modified to increment the `nonce` value until a valid hash is found. The `prepare_hash_data` method has been added to serialize the block's content for hashing. The `validate` method has been updated to compare the first `TARGET_LEN` characters of the hash with a target string. These changes enhance the mining process and ensure the integrity of the blockchain.