refactor: Add TARGET_LEN constant to Block
This commit adds a `TARGET_LEN` constant to the `Block` struct in order to specify the length of the target string for the proof of work algorithm. This constant will be used in the `validate` method to compare the first `TARGET_LEN` characters of the hash with the target string. This change enhances the flexibility and configurability of the proof of work implementation.