refactor: Update `Blockchain` struct to use sled::Db for storage
This commit refactors the `Blockchain` struct in the `blockchain.rs` file to use the `sled::Db` type for storage. It replaces the previous implementation that used a vector of blocks. The `Blockchain` struct now has a `current_hash` field to keep track of the current block's hash, and a `db` field to store the blocks using the `sled` crate. This change improves the efficiency and scalability of the blockchain implementation.